LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/effect-expected.txt


PASS effect is set correctly.
PASS Clearing and setting Animation.effect should update the computed style of the target element
